(goopt): move positional processing to a separate file
remove superseded functions
