+++ v2/goopt_internal.go
@@ -312,341 +312,6 @@ func (p *Parser) getArgumentInfoByID(id string) *FlagInfo {
 	return nil
 }
 
-type flagCache struct {
-	flags        map[string]map[string]*FlagInfo
-	needsValue   map[string]bool
-	isStandalone map[string]bool
-}
-
-func (p *Parser) buildFlagCache() *flagCache {
-	cache := &flagCache{
-		flags:        make(map[string]map[string]*FlagInfo),
-		needsValue:   make(map[string]bool),
-		isStandalone: make(map[string]bool),
-	}
-
-	for flag := p.acceptedFlags.Front(); flag != nil; flag = flag.Next() {
-		fv := flag.Value
-		longName := *flag.Key
-
-		// Extract the base flag name (before @)
-		baseName := splitPathFlag(longName)[0]
-
-		if _, exists := cache.flags[baseName]; !exists {
-			cache.flags[baseName] = make(map[string]*FlagInfo)
-		}
-
-		if shortName := fv.Argument.Short; shortName != "" {
-			if _, exists := cache.flags[shortName]; !exists {
-				cache.flags[shortName] = make(map[string]*FlagInfo)
-			}
-		}
-
-		cmdPath := ""
-		if fv.CommandPath != "" {
-			cmdPath = fv.CommandPath
-		}
-
-		cache.flags[baseName][cmdPath] = fv
-		if shortName := fv.Argument.Short; shortName != "" {
-			cache.flags[shortName][cmdPath] = fv
-			cache.isStandalone[shortName] = fv.Argument.TypeOf == types.Standalone
-			cache.needsValue[shortName] = fv.Argument.TypeOf != types.Standalone
-		}
-
-		// Store needsValue and isStandalone with the full key including command path
-		cache.isStandalone[longName] = fv.Argument.TypeOf == types.Standalone
-		cache.needsValue[longName] = fv.Argument.TypeOf != types.Standalone
-
-		// Also store with base name for global flags
-		if cmdPath == "" {
-			cache.isStandalone[baseName] = fv.Argument.TypeOf == types.Standalone
-			cache.needsValue[baseName] = fv.Argument.TypeOf != types.Standalone
-		}
-
-		// Don't treat positional arguments as flags that need values
-		if fv.Argument.Position != nil {
-			cache.needsValue[longName] = false
-			cache.needsValue[baseName] = false
-		}
-	}
-
-	return cache
-}
-
-func (p *Parser) setPositionalArguments(state parse.State) {
-	args := state.Args()
-	positional := make([]PositionalArgument, 0, len(args))
-	cache := p.buildFlagCache()
-
-	declaredPos := make([]struct {
-		key      string
-		flag     *FlagInfo
-		index    int
-		required bool
-	}, 0, p.acceptedFlags.Len())
-
-	for flag := p.acceptedFlags.Front(); flag != nil; flag = flag.Next() {
-		fv := flag.Value
-		if fv.Argument.Position != nil {
-			declaredPos = append(declaredPos, struct {
-				key      string
-				flag     *FlagInfo
-				index    int
-				required bool
-			}{
-				*flag.Key,
-				fv,
-				*fv.Argument.Position,
-				fv.Argument.Required,
-			})
-		}
-	}
-
-	if len(declaredPos) > 0 {
-		sort.SliceStable(declaredPos, func(i, j int) bool {
-			return declaredPos[i].index < declaredPos[j].index
-		})
-	}
-
-	skipNext := false
-	currentCmdPath := make([]string, 0, 3) // Pre-allocate for typical command depth
-	argPos := 0
-	executedCommands := make(map[string]bool) // Track which commands were encountered
-	executedCommands[""] = true               // Always check global positionals
-	for i, arg := range args {
-		if skipNext {
-			skipNext = false
-			continue
-		}
-
-		if p.isFlag(arg) {
-			name := strings.TrimLeft(arg, "-")
-
-			// Try to get canonical name from translation registry
-			canonicalName := name
-			if canonical, ok := p.translationRegistry.GetCanonicalFlagName(name, p.GetLanguage()); ok {
-				canonicalName = canonical
-			}
-
-			// Check if this flag needs a value
-			needsValue := false
-
-			// First check if this flag exists in the current command context
-			if len(currentCmdPath) > 0 {
-				if flagInfo, exists := cache.flags[canonicalName]; exists {
-					cmdPath := strings.Join(currentCmdPath, " ")
-					if cmdFlagInfo, cmdExists := flagInfo[cmdPath]; cmdExists {
-						// Found flag in command context
-						needsValue = cmdFlagInfo.Argument.TypeOf != types.Standalone && cmdFlagInfo.Argument.Position == nil
-					} else if globalFlagInfo, globalExists := flagInfo[""]; globalExists {
-						// Not in command context, but exists as global flag
-						needsValue = globalFlagInfo.Argument.TypeOf != types.Standalone && globalFlagInfo.Argument.Position == nil
-					}
-				}
-			} else {
-				// No command context, check global flags
-				if flagInfo, exists := cache.flags[canonicalName]; exists {
-					if globalFlagInfo, globalExists := flagInfo[""]; globalExists {
-						needsValue = globalFlagInfo.Argument.TypeOf != types.Standalone && globalFlagInfo.Argument.Position == nil
-					}
-				}
-			}
-
-			if needsValue {
-				skipNext = true
-			}
-			continue
-		}
-
-		// Special handling for standalone flags with boolean values
-		// If previous arg was a standalone flag and current arg is a valid boolean,
-		// skip it as it was consumed by the flag
-		if i > 0 && p.isFlag(args[i-1]) {
-			prevName := strings.TrimFunc(args[i-1], p.prefixFunc)
-
-			// Try to get canonical name from translation registry
-			canonicalPrevName := prevName
-			if canonical, ok := p.translationRegistry.GetCanonicalFlagName(prevName, p.GetLanguage()); ok {
-				canonicalPrevName = canonical
-			}
-
-			// Check if previous flag was standalone
-			isStandalone := false
-			if len(currentCmdPath) > 0 {
-				if flagInfo, exists := cache.flags[canonicalPrevName]; exists {
-					cmdPath := strings.Join(currentCmdPath, " ")
-					if cmdFlagInfo, cmdExists := flagInfo[cmdPath]; cmdExists {
-						isStandalone = cmdFlagInfo.Argument.TypeOf == types.Standalone
-					} else if globalFlagInfo, globalExists := flagInfo[""]; globalExists {
-						isStandalone = globalFlagInfo.Argument.TypeOf == types.Standalone
-					}
-				}
-			} else {
-				if flagInfo, exists := cache.flags[canonicalPrevName]; exists {
-					if globalFlagInfo, globalExists := flagInfo[""]; globalExists {
-						isStandalone = globalFlagInfo.Argument.TypeOf == types.Standalone
-					}
-				}
-			}
-
-			// If previous was standalone and current is a valid boolean, skip it
-			if isStandalone {
-				if _, err := strconv.ParseBool(arg); err == nil {
-					continue
-				}
-			}
-		}
-
-		// Check command path
-		isCmd := false
-		canonicalArg := arg
-		// Get canonical command name if it's a translated command
-		if canonical, ok := p.translationRegistry.GetCanonicalCommandPath(arg, p.GetLanguage()); ok {
-			canonicalArg = canonical
-		}
-
-		if len(currentCmdPath) == 0 {
-			if p.isCommand(arg) {
-				currentCmdPath = append(currentCmdPath, canonicalArg)
-				isCmd = true
-				argPos = 0 // Reset position counter for new command
-			}
-		} else {
-			switch {
-			case p.isCommand(strings.Join(append(currentCmdPath, arg), " ")):
-				currentCmdPath = append(currentCmdPath, canonicalArg)
-				isCmd = true
-			case p.isCommand(arg):
-				currentCmdPath = []string{canonicalArg}
-				isCmd = true
-				argPos = 0
-			default:
-			}
-		}
-
-		if isCmd {
-			// Mark this command path as executed
-			executedCommands[strings.Join(currentCmdPath, " ")] = true
-			continue
-		}
-
-		// Store the positional with its command context
-		pa := PositionalArgument{
-			Position: i,
-			Value:    arg,
-			Argument: nil,
-			ArgPos:   argPos,
-		}
-
-		// Find the matching declared positional for this command context
-		cmdPath := strings.Join(currentCmdPath, " ")
-		skipThisPositional := false
-		for _, decl := range declaredPos {
-			// Check if this declaration belongs to the current command path
-			if decl.flag.CommandPath == cmdPath && *decl.flag.Argument.Position == argPos {
-				lookup := buildPathFlag(decl.key, decl.flag.CommandPath)
-
-				// Check if this flag was already explicitly set via flag syntax
-				if _, alreadySet := p.options[lookup]; alreadySet {
-					// This positional slot was filled via explicit flag syntax
-					// Don't treat this value as a positional argument
-					skipThisPositional = true
-					continue
-				}
-
-				pa.Argument = decl.flag.Argument
-
-				// Run validators on positional argument
-				if len(decl.flag.Argument.Validators) > 0 {
-					for _, validator := range decl.flag.Argument.Validators {
-						if err := validator(arg); err != nil {
-							p.addError(errs.WrapOnce(err, errs.ErrProcessingFlag, lookup))
-						}
-					}
-				}
-
-				p.registerFlagValue(lookup, arg, arg)
-				p.options[lookup] = arg
-				if err := p.setBoundVariable(arg, lookup); err != nil {
-					p.addError(errs.ErrSettingBoundValue.WithArgs(lookup).Wrap(err))
-				}
-				break
-			}
-		}
-
-		if !skipThisPositional {
-			positional = append(positional, pa)
-		}
-		argPos++
-	}
-
-	// Check for missing required positionals and apply defaults
-	// Since we already matched positionals during collection, we only need to check for missing ones
-	for _, decl := range declaredPos {
-		// Only check positionals for commands that were actually executed
-		if !executedCommands[decl.flag.CommandPath] {
-			continue
-		}
-
-		// Check if this positional was provided
-		found := false
-		for _, pos := range positional {
-			if pos.Argument == decl.flag.Argument {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			// Check if a flag was provided that takes precedence
-			flagFromArg := p.flagOrShortFlag(decl.key, decl.flag.CommandPath)
-			if p.HasRawFlag(flagFromArg) {
-				continue
-			}
-
-			if decl.flag.Argument.DefaultValue != "" {
-				// Apply default value
-				lookup := buildPathFlag(decl.key, decl.flag.CommandPath)
-				p.registerFlagValue(lookup, decl.flag.Argument.DefaultValue, decl.flag.Argument.DefaultValue)
-				p.options[lookup] = decl.flag.Argument.DefaultValue
-				if err := p.setBoundVariable(decl.flag.Argument.DefaultValue, lookup); err != nil {
-					p.addError(errs.ErrSettingBoundValue.WithArgs(lookup).Wrap(err))
-				}
-
-				// Add a positional argument entry for the default value
-				positional = append(positional, PositionalArgument{
-					Position: decl.index, // Use the declared position
-					Value:    decl.flag.Argument.DefaultValue,
-					Argument: decl.flag.Argument,
-					ArgPos:   decl.index,
-				})
-			} else if decl.required {
-				// Missing required positional
-				p.addError(errs.ErrRequiredPositionalFlag.WithArgs(decl.key, decl.index))
-			}
-		}
-	}
-
-	newResult := make([]PositionalArgument, 0, len(positional))
-
-	for i := range positional {
-		// Keep positions that either:
-		// 1. Have a non-empty value
-		// 2. Are unbound (Argument == nil) and have any value
-		if positional[i].Value != "" {
-			newResult = append(newResult, positional[i])
-		}
-	}
-
-	// Sort by position to maintain order
-	sort.Slice(newResult, func(i, j int) bool {
-		return newResult[i].Position < newResult[j].Position
-	})
-
-	p.positionalArgs = newResult
-}
-
 func (p *Parser) evalExecOnParse(lastCommandPath string) string {
 	if p.callbackOnParse {
 		err := p.ExecuteCommand()
